Implement Firestore db in Doctor and Nurse pages
Doctors and Nurses are now able to create and store new resident information in the Firestore Database.

Added fakes information for Employees and Residents so they don't mix data with each other. Next iteration we need to have a check for duplicate employee and resident data.

Moved imports to register.js